requirements.txt
django-heroku==0.3.1
Django==2.2.1
djangorestframework==3.9.4